implement profile screen, refactor more stack
